problem 100 & changed return types from int to long for more flexibility

+++ ProjectEuler/src/com/brady/euler/Problem.java
@@ -5,7 +5,8 @@ import java.util.Scanner;
 
 public abstract class Problem {
 	
-	public int solve() {
+	
+	public long solve() {
 		throw new UnsupportedOperationException("Not implemented");
 	}
